done some session thing
the createSeesionBitTable creates a table to check if seesion had
started, if not, students cannot login.

login.php is updated to check if session for student had started if not,
do not let login.

session.php is editted to update the sessionBit table, setting it to '1'
when start and '0' when end.